requirement.txt
django==4.1
djangorestframework
requests
django-cors-headers
djongo
sqlparse==0.2.4
pymongo==3.12.3
psycopg2
django[bcrypt]
django-cors-headers
djangorestframework_simplejwt
dj-rest-auth
dj-rest-auth[with_social]
social-auth-app-django
six
django-filter==23.1
